requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
translate==3.6.1
pandas==1.3.5
scikit-learn==1.3.0
numpy==1.24.3
torch==2.0.1
torchvision==0.17.0
pillow==9.5.0
tqdm==4.65.0
matplotlib==3.7.1
certifi>=2023.7.22
requests==2.26.0
transformers==4.30.2 